# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AdminResponse:
    properties:
      data: {}
      message:
        type: string
    type: object
...
  dto.AttendancePeriodData:
    properties:
      end_date:
        type: string
      id:
        type: integer
      start_date:
        type: string
    type: object
  dto.AttendancePeriodRequest:
    properties:
      end_date:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - start_date
    type: object
  dto.AttendancePeriodResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.AttendancePeriodData'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_record:
        type: integer
    type: object
...
  dto.AuditLogData:
    properties:
      action:
        type: string
      full_name:
        type: string
      id:
        type: integer
      ip_address:
        type: string
      record_id:
        type: integer
      request_id:
        type: string
      table_name:
        type: string
    type: object
  dto.AuditLogResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.AuditLogData'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_record:
        type: integer
    type: object
  dto.AuthRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  dto.AuthResponse:
    properties:
      expire:
        type: string
      token:
        type: string
      type:
        type: string
    type: object
...
  dto.PayrollRequest:
    properties:
      period_id:
        type: integer
    type: object
  dto.PayrollsResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.PayrollsResponseData'
        type: array
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_record:
        type: integer
    type: object
  dto.PayrollsResponseData:
    properties:
      end_date:
        type: string
      id:
        type: integer
      processed_at:
        type: string
      start_date:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API untuk manajemen payslip.
  title: Payslip API
  version: "1.0"
paths:
  /admin/attendance-periods:
    post:
      consumes:
      - application/json
      description: Create a new attendance period
      parameters:
      - description: attendance period create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AttendancePeriodRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AdminResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create attendance period
      tags:
      - admin
  /admin/attendance-periods/all:
    get:
      consumes:
      - application/json
      description: Retrieves a list of attendance period with pagination.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AttendancePeriodResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List of attendance period
      tags:
      - admin
  /admin/payroll:
    post:
      consumes:
      - application/json
      description: Create a new payroll
      parameters:
      - description: period create payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.PayrollRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AdminResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create payroll
      tags:
      - admin
  /admin/payroll/{id}/process:
    put:
      consumes:
      - application/json
      description: Run a new payroll
      parameters:
      - description: Payroll ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AdminResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Run payroll
      tags:
      - admin
  /admin/payroll/{id}/summary:
    get:
      consumes:
      - application/json
      description: Summary a payroll
      parameters:
      - description: Payroll ID
        in: path
        name: id
        required: true
        type: integer
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AdminResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Summary payroll
      tags:
      - admin
  /admin/payroll/all:
    get:
      consumes:
      - application/json
      description: Retrieves a list of payroll with pagination.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.PayrollsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List of payroll
      tags:
      - admin
  /auditlogs/all:
    get:
      consumes:
      - application/json
      description: Retrieves a list of audit logs with pagination.
      parameters:
      - description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Items per page (default: 20)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuditLogResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List of audit logs
      tags:
      - auditlogs
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AuthRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"